requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
weasyprint==60.1
python-multipart==0.0.6
resend==0.6.0
python-dotenv==1.0.0
markdown==3.5.1 